browser-automation-api/requirements.txt
langchain-openai
browser-use
python-dotenv
asyncio
requests
flask
flask-cors
pytest
pytest-asyncio